requirements.txt
colorama
numpy
openai
protobuf
psutil
pyshark
scikit-learn
pandas
torch
seaborn
httpx
openai
pyshark
python-dotenv
boto3
anthropic
pyyaml
mistralai
transformers
peft
groq
pandas
tqdm
matplotlib
lightgbm